feat(Reservation): 관리자 예약 삭제 API 예외 처리 구현

+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -76,6 +76,10 @@ public class ReservationService {
     }
 
     public void deleteReservationById(Long id) {
+        if (!reservationRepository.existsReservationById(id)) {
+            throw new ReservationException(ReservationErrorType.RESERVATION_NOT_FOUND);
+        }
+
         reservationRepository.deleteReservationById(id);
     }
 }
